requirements.txt
django==1.11.7
django-extensions==1.9.7
wechatpy==1.6.1
mysqlclient==1.3.12
beautifulsoup4==4.6.0
html5lib==0.999999999
Pillow==2.6.1
django-celery==3.2.2
django-braces==1.8.1
django-filter==1.1.0
django-taggit==0.22.2
django-autocomplete-light==3.2.10
djangorestframework==3.6.4
djangorestframework-csv==1.3.0
drf-extensions==0.3.1
easy-thumbnails==2.5
pytz==2017.3
django-redis==4.8.0
redis==2.10.6
django-dbsettings==0.10.0
django-guardian==1.4.9
django-js-reverse==0.8.1
python-dateutil
Wand==0.4.3
enum34
pyenchant
weixin-python
xhtml2pdf==0.0.6
pyOpenSSL==17.3.0
pathlib==1.0.1
jsonfield==2.0.2
requests==2.11.1
urlfetch==1.0.2
django-stdimage==2.4.2
pypinyin==0.27
aliyun-python-sdk-core-v3==2.8.6
numpy==1.13.3
lxml==4.1.1
django-rest-swagger
wagtail==1.13.1
django-environ==0.4.4
django-tinymce4-lite==1.6.0
